loop i long

+++ PI.c
@@ -9,7 +9,7 @@ int main(int argc, char *argv[])
 
   unsigned int seed = omp_get_wtime();
 
-  for (int i = 0; i < num_steps; i++) {
+  for (long i = 0; i < num_steps; i++) {
       double x = ((double)rand_r(&seed) / RAND_MAX);
       double y = ((double)rand_r(&seed) / RAND_MAX);
